# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FASE 2 API",
        "description": "FASE 2 Swagger OpenApi",
        "contact": {
            "email": "eaguilar.arrezola@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/agents": {
            "get": {
                "tags": [
                    "agents"
                ],
                "summary": "Get all agents",
                "operationId": "c3472e4b464378a04737b2b69e244fda",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "agents"
                ],
                "summary": "Add agent",
                "operationId": "9970c27ad1c69325505ae9c90bf30222",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/agents/{id}": {
            "get": {
                "tags": [
                    "agents"
                ],
                "summary": "Get agent",
                "operationId": "b2d65243e40b214088f810608a0e95b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "agents"
                ],
                "summary": "Update agent",
                "operationId": "e910d56058655e786c26d8edffa93ed6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "agents"
                ],
                "summary": "Delete agent",
                "operationId": "0f375894d997d30ec163b6159e55deb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/cat_packages/paginate": {
            "post": {
                "tags": [
                    "cat_packages"
                ],
                "summary": "Get cat_packages per paginate",
                "operationId": "49bd0eb2e70dbc79be3c14a8689e31a0",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/cat_packages": {
            "get": {
                "tags": [
                    "cat_packages"
                ],
                "summary": "Get all cat_packages",
                "operationId": "d47dbb79b2acc4f5c8bdf9df8bf59761",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/cat_packages/{id}": {
            "get": {
                "tags": [
                    "cat_packages"
                ],
                "summary": "Get cat_packages",
                "operationId": "2f938a34963fdb406b7299168c506d33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "cat_packages"
                ],
                "summary": "Update cat_packages",
                "operationId": "20bc0919862b074a9662e59201f67274",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "cat_packages"
                ],
                "summary": "Add cat_packages",
                "operationId": "482dc060b9083bbcf16a3b9880395c72",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "cat_packages"
                ],
                "summary": "Delete cat_packages",
                "operationId": "21b42acc8b1ef117958af4fdfc39fb83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}